backend/requirements.txt
fastapi==0.111.0
bcrypt==4.0.1
passlib==1.7.4
pydantic==2.7.1
python-dotenv==1.0.1
python_jose==3.3.0
Requests==2.31.0
SQLAlchemy==2.0.29
